feat: update

+++ DJTest/Configuration/Common/version.commit.sh
@@ -17,5 +17,11 @@ file_config='Common.xcconfig'
 previous_build_number=$(awk -F "=" '/'$key_version' = / {print $2}' $file_config | tr -d ' ')
 echo "-->build version[$CONFIGURATION]: $previous_build_number -> $build_version"
 
+# 每 20 个 commit 变更一次版本号
+mod=$(($number_of_commits % 20))
+if [ $mod -eq 0 ]; then
 sed -i -e "/$key_version =/ s/= .*/= $build_version/" $file_config
 rm -f $file_config-e
+else
+echo "skipping[$mod]..."
+fi
